Stop it and tidy up
