+++ wasm-calc10/src/Calc/Typecheck/Patterns.hs
@@ -1,0 +1,7 @@
+module Calc.Typecheck.Patterns (
+  module Calc.Typecheck.Patterns.Annihilate,
+  module Calc.Typecheck.Patterns.Validate, module Calc.Typecheck.Patterns.Generate) where
+
+import Calc.Typecheck.Patterns.Annihilate
+import Calc.Typecheck.Patterns.Generate
+import Calc.Typecheck.Patterns.Validate
