+++ wasm-calc10/test/Main.hs
@@ -5,7 +5,7 @@ import Test.Hspec
 import qualified Test.Linearity.LinearitySpec
 import qualified Test.Parser.ParserSpec
 import qualified Test.PrettyPrint.PrettyPrintSpec
-import qualified Test.Typecheck.ExhaustivenessSpec
+import qualified Test.Typecheck.PatternsSpec
 import qualified Test.Typecheck.TypecheckSpec
 import qualified Test.Wasm.FromWasmSpec
 import qualified Test.Wasm.WasmSpec
@@ -17,6 +17,6 @@ main = hspec $ do
   Test.PrettyPrint.PrettyPrintSpec.spec
   Test.Linearity.LinearitySpec.spec
   Test.Typecheck.TypecheckSpec.spec
-  Test.Typecheck.ExhaustivenessSpec.spec
+  Test.Typecheck.PatternsSpec.spec
   Test.Wasm.WasmSpec.spec
   Test.Wasm.FromWasmSpec.spec
